pyproject.toml
[project]
name = "pixi-devenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ESSS", email = "dev@esss.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyserde[toml]>=0.25.0",
    "tomlkit>=0.13.3",
    "typer>=0.17.3",
]

[project.scripts]
pixi-devenv = "pixi_devenv.cli:app"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-regressions>=2.8.3",
    "ruff>=0.12.12",
]
